[ENH]Refacroring: re organised app structure to support external module
