+++ config/config.ini
@@ -7,3 +7,6 @@ host="localhost"
 [app_conf]
 link=""
 params=""
+
+[autoload]
+vendor=false
