+++ global/cors.php
