+++ index.php
@@ -1,5 +1,4 @@
 <?php
-    require "vendor/autoload.php";
-    require_once 'config/init.php';
+    require_once 'global/init.php';
     require_once 'route/index.php';
 ?>
